+++ packages/build-utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @vercel/build-utils
 
+## 13.8.2
+
+### Patch Changes
+
+- [services] allow services to share builder source ([#15631](https://github.com/vercel/vercel/pull/15631))
+
+- Updated dependencies [[`8e8110d2eca5832e109f5efb64b192690100927d`](https://github.com/vercel/vercel/commit/8e8110d2eca5832e109f5efb64b192690100927d)]:
+  - @vercel/python-analysis@0.10.1
+
 ## 13.8.1
 
 ### Patch Changes